move schedule utils
