+++ app/component/routepage/schedule/scheduleTripsUtils.js
